rinaperf: fix compilation problem on 32-bit systems

+++ user/rinaperf.c
@@ -47,7 +47,8 @@ client_test_config(struct rinaperf *rp)
         if (ret < 0) {
             perror("write(buf)");
         } else {
-            printf("%s: partial write %d/%lu\n", __func__, ret, sizeof(cfg));
+            printf("%s: partial write %d/%lu\n", __func__, ret,
+                    (unsigned long int)sizeof(cfg));
         }
         return -1;
     }
@@ -66,7 +67,8 @@ server_test_config(struct rinaperf *rp)
         if (ret < 0) {
             perror("read(buf");
         } else {
-            printf("%s: partial write %d/%lu\n", __func__, ret, sizeof(cfg));
+            printf("%s: partial write %d/%lu\n", __func__, ret,
+                    (unsigned long int)sizeof(cfg));
         }
         return -1;
     }
